requirements.txt
# OCR / PDF
opencv-python==4.12.0.88
pytesseract==0.3.13
pdfplumber==0.11.8
camelot-py[cv]==1.0.9

# Modelos LLM / Embeddings
transformers==4.57.1
torch==2.9.1
huggingface-hub==0.36.0
tokenizers==0.22.1
safetensors==0.6.2
faiss-cpu==1.13.0
sentence-transformers==5.1.2

# PDFs / Plantillas
Jinja2==3.0.3
WeasyPrint==60.0
reportlab==4.4.5
Pillow==12.0.0

# Interfaz / App
streamlit==1.51.0
